refactor: 데이터 형식 double로 변경

+++ src/main/java/com/deepnyangning/capstonebe/domain/access/dto/AccessLogResponse.java
@@ -22,5 +22,5 @@ public class AccessLogResponse {
 
     private LocalDateTime accessTime;
 
-    private float similarity;
+    private double similarity;
 }
